requirements.txt
numpy~=1.21.6
fastapi~=0.92.0
tensorflow~=2.11.0
Pillow~=9.4.0
uvicorn~=0.20.0

opencv-python~=4.7.0.72